backend/requirements.txt
# webapp-boilerplate backend dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlmodel>=0.0.22
sqlalchemy>=2.0.35
aiosqlite>=0.20.0
alembic>=1.14.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0
pydantic[email]
email-validator>=2.0.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.28.0

# AI integrations for multilingual support
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
Faker>=33.0.0
